pyproject.toml
[project]
name = "a2a-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.25",
    "uvicorn>=0.43.0",
]
